feat(archives & organizations update): archives and trash System. Organization, roles, superadmin
Factories are renamed Repositories. Documents can now be archived. Archived documents will not be
sent with getAll. Removing a document will not delete it. Removed documents can be accessed from
getRemoved, and permanently detroyed only if they were first removed. Organizations is a new
mandatory repository. User are associated with one or multiple organizations, and so are their
documents. A user can only access the documents associated with their origanizations. Roles are
natively included as a numbers array. Superadmin role, number 1, is allowed everything.

BREAKING CHANGE: New permission function are required with a model declaration